+++ aeron-client/src/main/cpp/concurrent/logbuffer/BufferClaim.h
@@ -39,14 +39,14 @@ public:
     /// @cond HIDDEN_SYMBOLS
     inline void wrap(std::uint8_t *buffer, util::index_t length)
     {
-        m_buffer.wrap(buffer, length);
+        m_buffer.wrap(buffer, static_cast<size_t>(length));
     }
     /// @endcond
 
     /// @cond HIDDEN_SYMBOLS
     inline void wrap(AtomicBuffer& buffer, util::index_t offset, util::index_t length)
     {
-        m_buffer.wrap(buffer.buffer() + offset, length);
+        m_buffer.wrap(buffer.buffer() + offset, static_cast<size_t>(length));
     }
     /// @endcond
